fix issue 15 on supporting config feature

+++ Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.IO;
 using McMaster.Extensions.CommandLineUtils;
+using Newtonsoft.Json.Linq;
+
 namespace shinny_ssg
 {
     static class Globals
@@ -24,8 +26,10 @@ namespace shinny_ssg
                     //help option
                     app.HelpOption();
                     app.VersionOption("-v|--version", "0.1", "Shinny SSG 0.1");
-                    var inputFileOption = app.Option<string>("-i|--input", "Input file/folder to convert source file to HTML", CommandOptionType.SingleValue)
-                                           .IsRequired();
+                    //config Option
+                    var configOption = app.Option<string>("-c|--config", "Configuration JSON File", CommandOptionType.SingleValue);
+                    // input file now is not required 
+                    var inputFileOption = app.Option<string>("-i|--input", "Input file/folder to convert source file to HTML", CommandOptionType.SingleValue);
                     var outputOption = app.Option<string>("-o|--output", "Output folder for converted file/files", CommandOptionType.SingleValue);
                     var cssOption = app.Option<string>("--stylesheet| -s", "Style Sheet for the converted HTML file", CommandOptionType.SingleValue);
                     //language option
@@ -34,27 +38,51 @@ namespace shinny_ssg
 
                     app.OnExecute(() =>
                     {
-                        var inputname = inputFileOption.Value();
-                        Globals.cssUrl = cssOption.HasValue() ? cssOption.Value() : null;
-                        Globals.langAtr = langOption.HasValue() ? langOption.Value() : null;
-                        if (outputOption.HasValue() && Directory.Exists(outputOption.Value()))
+                        var configName = configOption.HasValue() ? configOption.Value() : null;
+                        var inputValue = inputFileOption.HasValue() ? inputFileOption.Value() : null;
+
+                        // If config worked 
+
+                        var inputname = "";
+                        if (configName != null && File.Exists(configName) && (configName.EndsWith(".json")))
                         {
-                            destination = outputOption.Value();
+                            //start working with config File here
+                            string jsonString = File.ReadAllText(configName);
+                            JObject jObj = JObject.Parse(jsonString);
+                            inputname = jObj.ContainsKey("input") ? (string)jObj["input"] : "";
+                            Globals.cssUrl = jObj.ContainsKey("stylesheet") ? (string)jObj["stylesheet"] : null;
+                            Globals.langAtr = jObj.ContainsKey("lang") ? (string)jObj["lang"] : null;
+                            destination = jObj.ContainsKey("lang") ? (string)jObj["output"] : default;
+                        }
+                        else if (inputValue != null)
+                        {
+                            Globals.cssUrl = cssOption.HasValue() ? cssOption.Value() : null;
+                            Globals.langAtr = langOption.HasValue() ? langOption.Value() : null;
+                            if (outputOption.HasValue() && Directory.Exists(outputOption.Value()))
+                            {
+                                destination = outputOption.Value();
+                            }
+                            else
+                            {
+                                //It will delete all file even though the read or write process fail
+                                System.IO.DirectoryInfo di = new DirectoryInfo(destination);
+                                foreach (FileInfo file in di.GetFiles(""))
+                                {
+                                    file.Delete();
+                                }
+                                foreach (DirectoryInfo dir in di.GetDirectories())
+                                {
+                                    dir.Delete(true);
+                                }
+
+                            }
+                            inputname = inputValue;
                         }
                         else
                         {
-                            //It will delete all file even though the read or write process fail
-                            System.IO.DirectoryInfo di = new DirectoryInfo(destination);
-                            foreach (FileInfo file in di.GetFiles(""))
-                            {
-                                file.Delete();
-                            }
-                            foreach (DirectoryInfo dir in di.GetDirectories())
-                            {
-                                dir.Delete(true);
-                            }
-
+                            throw new Exception("Either Input or Config File must have a valid value ");
                         }
+
 
                         if (File.Exists(inputname) && (inputname.EndsWith(".txt") || inputname.EndsWith(".md")))
                         {
@@ -73,12 +101,13 @@ namespace shinny_ssg
                         {
                             var f = new Subfolder();
                             f.CreateFolder(inputname, destination);
-
                         }
                         else
                         {
                             Console.WriteLine("Input Path is not valid.");
                         }
+
+
 
                     });
 
@@ -92,7 +121,6 @@ namespace shinny_ssg
             }
             catch (Exception ex)
             {
-                Console.Error.WriteLine("There is an error with the transfer process");
                 Console.Error.WriteLine(ex.Message);
                 return -1;
             }
